refactor: move log build plugin to its own file
