+++ scripts/build.mts
@@ -2,10 +2,10 @@ import { createContext } from "@marshift/argus";
 import esbuild from "esbuild";
 import { rmSync } from "fs";
 import path from "path";
-import { logBuildPlugin } from "src/util.mjs";
 import { fileURLToPath } from "url";
 import intlPlugin from "./build-plugins/intl-loader.mjs";
 import intlTypeGeneratorPlugin from "./build-plugins/intl-type-generator.mjs";
+import logBuildPlugin from "./build-plugins/log-build.mjs";
 import preBundlePlugin from "./build-plugins/pre-bundle.mjs";
 
 const NODE_VERSION = "20";
